added constants folder and renamed errors to statusCodes
