+++ ipfs/ipfs.go
@@ -24,7 +24,7 @@ import (
 	"github.com/ipfs/go-ipfs/plugin/loader" // This package is needed so that all the preloaded plugins are loaded automatically
 	"github.com/ipfs/go-ipfs/repo/fsrepo"
 	"github.com/libp2p/go-libp2p-core/peer"
-	status "github.com/sonr-io/core/errors"
+	status "github.com/sonr-io/core/constants/status"
 )
 
 var FlagExp = flag.Bool("experimental", false, "enable experimental features")
